routes
route de reservas con filtros

+++ routes/reserva.js
@@ -95,42 +95,59 @@ router.get('/activas', async(req, res, next) => {
                 });
             }
         }).catch(error => {
-        res.status(400).json({
-            status: 0,
-            statusCode: 'database/error',
-            description: error.toString()
+            res.status(400).json({
+                status: 0,
+                statusCode: 'database/error',
+                description: error.toString()
+            });
         });
-    });
 });
 
 //PUT-UPDATE libera la mesa, es decir estado de la reserva igual a cero(false)
-router.patch('/libera/:mesaNumero', async (req, res, next) => {
-    const fecha_inicio_reserva = req.body['fecha_inicio_reserva'];
-    const usuarioRut = req.body['usuarioRut'];
-    const mesaNumero = req.params.mesaNumero;
+router.patch('/libera/', async (req, res, next) => {
+    
+    const rut = req.body['rut'];
+    const password = req.body['password'];
+    const id = req.body['id']; //id de reserva
 
-    if (mesaNumero && usuarioRut && fecha_inicio_reserva) {
-        models.reserva.findOne({
+    if (rut && password && id) {
+        models.usuario.findOne({
             where: {
-                mesaNumero: mesaNumero,
-               fecha_inicio_reserva: fecha_inicio_reserva,
-                usuarioRut : usuarioRut
+                rut: rut,
+                password: password,
+                tipo_usuario: 'Admin'
             }
-        }).then(reserva => {
-            if (reserva) {
-                reserva.updateAttributes({
-                    estado:false
+        }).then(usuario => {
+            if (usuario) {
+                models.reserva.findOne({
+                    where:{
+                        id:id
+                    }
+
+                }).then(reserva => {
+
+                    reserva.updateAttributes({
+                        estado:false
+                    });
+                    res.json({
+                        status: 1,
+                        statusCode: 'reserva/liberada',
+                        data: reserva.toJSON()
+                    });
+                }).catch(error => {
+                    res.status(400).json({
+                        status: 0,
+                        statusCode: 'database/error',
+                        description: error.toString()
+                    });
                 });
-                res.json({
-                    status: 1,
-                    statusCode: 'reserva/found',
-                    data: reserva.toJSON()
-                });
+
+
             } else {
                 res.status(400).json({
                     status: 0,
-                    statusCode: 'reserva/not-found',
-                    description: 'The reserva was not found with the numero'
+                    statusCode: 'usuario/not-found',
+                    description: 'No se encontro usuario con datos'
                 });
             }
         }).catch(error => {
@@ -143,32 +160,51 @@ router.patch('/libera/:mesaNumero', async (req, res, next) => {
     } else {
         res.status(400).json({
             status: 0,
-            statusCode: 'reserva/wrong-atts',
-            description: 'Check the atribs!'
+            statusCode: 'usuario/wrong-keys-or-id',
+            description: 'Check the keys and id!'
         });
     }
 });
 
 //DELETE-DELETE por id de reserva
-router.delete('/delete/:id', async(req, res, next) => {
-    const id = req.params.id;
-    if (id) {
-        models.reserva.destroy({
+router.delete('/delete/', async(req, res, next) => {
+    
+    const rut = req.body['rut'];
+    const password = req.body['password'];
+    const id = req.body['id'];
+
+    if (rut && password && id) {
+        models.usuario.findOne({
             where: {
-                id: id
+                rut: rut,
+                password: password,
+                tipo_usuario: 'Admin'
             }
-        }).then(reserva => {
-            if (reserva) {
-                res.json({
-                    status: 1,
-                    statusCode: 'reserva deleted',
-                    //data: reserva.toJSON()
+        }).then(usuario => {
+            if (usuario) {
+                models.reserva.destroy({
+                    where:{
+                        id:id
+                    }
+                }).then(reserva => {
+                    res.json({
+                        status: 1,
+                        statusCode: 'reserva/eliminada',
+                    });
+                }).catch(error => {
+                    res.status(400).json({
+                        status: 0,
+                        statusCode: 'database/error',
+                        description: error.toString()
+                    });
                 });
+
+
             } else {
                 res.status(400).json({
                     status: 0,
-                    statusCode: 'reserva no borrado',
-                    description: 'The reserva was not found with the numero'
+                    statusCode: 'usuario/not-found',
+                    description: 'No se encontro usuario con datos'
                 });
             }
         }).catch(error => {
@@ -181,8 +217,8 @@ router.delete('/delete/:id', async(req, res, next) => {
     } else {
         res.status(400).json({
             status: 0,
-            statusCode: 'user/wrong-id',
-            description: 'Check the id!'
+            statusCode: 'usuario/wrong-keys-or-id',
+            description: 'Check the keys and id!'
         });
     }
 });
